test: vms are restarted after irrecoverable post-copy migration
