+++ tests/migration/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "framework.go",
         "migration.go",
         "migration_policy.go",
+        "postcopy.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/migration",
     visibility = ["//visibility:public"],
@@ -60,6 +61,8 @@ go_library(
         "//vendor/github.com/google/goexpect:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/github.com/onsi/gomega/gstruct:go_default_library",
+        "//vendor/github.com/onsi/gomega/types:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/api/policy/v1:go_default_library",
         "//vendor/k8s.io/api/policy/v1beta1:go_default_library",
